src/requirements.txt
streamlit
requests
opencv-python
Pillow
python-dotenv
openai
pandas
azure-identity
azure-storage-blob
setuptools 
wheel 
pygame
ipywidgets